feat: hook up team page bot buttons
