+++ shared/teams/team/index.tsx
@@ -39,6 +39,7 @@ class Team extends React.Component<Props> {
       case 'header':
       case 'member':
       case 'bot':
+      case 'bot-add':
       case 'invites-invite':
       case 'invites-request':
       case 'invites-divider':
