+++ shared/teams/team/rows/bot-row/bot/container.tsx
@@ -1,12 +1,14 @@
-import * as Constants from '../../../../constants/teams'
-import * as Chat2Gen from '../../../../actions/chat2-gen'
-import * as Tracker2Gen from '../../../../actions/tracker2-gen'
-import * as ProfileGen from '../../../../actions/profile-gen'
-import * as Types from '../../../../constants/types/teams'
-import * as RPCTypes from '../../../../constants/types/rpc-gen'
+import * as Constants from '../../../../../constants/teams'
+import * as Chat2Constants from '../../../../../constants/chat2'
+import * as Chat2Gen from '../../../../../actions/chat2-gen'
+import * as Tracker2Gen from '../../../../../actions/tracker2-gen'
+import * as ProfileGen from '../../../../../actions/profile-gen'
+import * as Types from '../../../../../constants/types/teams'
+import * as RPCTypes from '../../../../../constants/types/rpc-gen'
 import {TeamBotRow} from './'
-import * as RouteTreeGen from '../../../../actions/route-tree-gen'
-import {connect, isMobile} from '../../../../util/container'
+import * as RouteTreeGen from '../../../../../actions/route-tree-gen'
+import {connect, isMobile} from '../../../../../util/container'
+import * as Chat2Types from '../../../../../constants/types/chat2'
 
 type OwnProps = {
   teamID: Types.TeamID
@@ -16,6 +18,8 @@ type OwnProps = {
 const blankInfo = Constants.initialMemberInfo
 
 type DispatchProps = {
+  _onEdit: (conversationIDKey: Chat2Types.ConversationIDKey, username: string) => void
+  _onRemove: (conversationIDKey: Chat2Types.ConversationIDKey, username: string) => void
   _onShowTracker: (username: string) => void
   onChat: () => void
   onClick: () => void
@@ -24,6 +28,7 @@ type DispatchProps = {
 export default connect(
   (state, {teamID, username}: OwnProps) => {
     const teamDetails = Constants.getTeamDetails(state, teamID)
+    const generalChannel = Chat2Constants.getChannelForTeam(state, teamDetails.teamname, 'general')
     const {members: map = new Map<string, Types.MemberInfo>(), teamname} = teamDetails
     const info: Types.MemberInfo = map.get(username) || blankInfo
     const bot: RPCTypes.FeaturedBot = state.chat2.featuredBotsMap.get(username) ?? {
@@ -37,6 +42,7 @@ export default connect(
 
     return {
       ...bot,
+      _convID: generalChannel.conversationIDKey,
       ownerTeam: bot.ownerTeam || undefined,
       ownerUser: bot.ownerUser || undefined,
       roleType: info.type,
@@ -47,6 +53,23 @@ export default connect(
     }
   },
   (dispatch, ownProps: OwnProps): DispatchProps => ({
+    _onEdit: (conversationIDKey: Chat2Types.ConversationIDKey, username: string) =>
+      dispatch(
+        RouteTreeGen.createNavigateAppend({
+          path: [
+            {
+              props: {
+                botUsername: username,
+                conversationIDKey,
+                namespace: 'chat2',
+              },
+              selected: 'chatInstallBot',
+            },
+          ],
+        })
+      ),
+    _onRemove: (conversationIDKey: Chat2Types.ConversationIDKey, username: string) =>
+      dispatch(Chat2Gen.createRemoveBotMember({conversationIDKey, username})),
     _onShowTracker: (username: string) => {
       if (isMobile) {
         dispatch(ProfileGen.createShowUserProfile({username}))
@@ -67,6 +90,8 @@ export default connect(
     botAlias: stateProps.botAlias,
     description: stateProps.description,
     onClick: dispatchProps.onClick,
+    onEdit: () => dispatchProps._onEdit(stateProps._convID, ownProps.username),
+    onRemove: () => dispatchProps._onRemove(stateProps._convID, ownProps.username),
     onShowTracker: () => dispatchProps._onShowTracker(ownProps.username),
     ownerTeam: stateProps.ownerTeam,
     ownerUser: stateProps.ownerUser,
