+++ shared/teams/team/rows/index.tsx
@@ -4,7 +4,7 @@ import {getOrderedMemberArray, sortInvites, getOrderedBotsArray} from './helpers
 type HeaderRow = {key: string; type: 'header'}
 type TabsRow = {key: string; type: 'tabs'}
 type MemberRow = {key: string; username: string; type: 'member'}
-type BotRow = {key: string; username: string; type: 'bot'}
+type BotRow = {key: string; username: string; type: 'bot'} | {key: string; type: 'bot-add'}
 type InviteRow =
   | {key: string; label: string; type: 'invites-divider'}
   | {key: string; username: string; type: 'invites-request'}
@@ -52,6 +52,7 @@ const makeRows = (
         // loading
         rows.push({key: 'loading', type: 'loading'})
       }
+      rows.push({key: 'bot:install-more', type: 'bot-add'})
       break
     case 'invites': {
       const {invites, requests} = details
